pyproject.toml
[tool.poetry]
name = "repl_python3_Filterhaj"
version = "0.1.0"
description = ""
authors = ["neeltron <<>>"]

[tool.poetry.dependencies]
python = "^3.8"
discord = "^1.7.3"
Flask = "^2.0.1"
Pillow = "^8.3.1"
youtube_dl = "^2021.6.6"
replit = "^3.2.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
